requirements.txt
pandas
requests
beautifulsoup4
selenium
geopandas
matplotlib
seaborn
jupyter
python-dotenv
rapidfuzz
sgp4
tenacity
